Fix pointer application after json roundtrip
The rfc6901 pointer string representation does explicitely say that application sementaics of a pointer is depended on the target type.
The pointer representation does not contain information whether the target is an object or an array in the case of a number.
To workaround this we are more lenient when applying a parsed Pointer.